requirements.txt
# python==3.9.20

absl-py==2.1.0
accelerate==0.26.0
addict==2.4.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
async-timeout==4.0.3
attrs==24.2.0
bbox-visualizer==0.1.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cmake==3.31.0.1
colorama==0.4.6
contourpy==1.3.0
crcmod==1.7
cryptography==43.0.3
cycler==0.12.1
datasets==3.1.0
diffusers==0.31.0
dill==0.3.8
docker-pycreds==0.4.0
filelock==3.14.0
fonttools==4.54.1
frozenlist==1.5.0
fsspec==2024.9.0
ftfy==6.3.1
gitdb==4.0.11
GitPython==3.1.43
grpcio==1.67.1
huggingface-hub==0.26.2
idna==3.10
imageio==2.36.0
importlib_metadata==8.5.0
importlib_resources==6.4.5
Jinja2==3.1.4
jmespath==0.10.0
kiwisolver==1.4.7
lazy_loader==0.4
lit==18.1.8
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
mdurl==0.1.2
mmcv-full==1.7.2
mmdet==2.26.0
model-index==0.1.11
modelcards==0.1.6
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.2.1
ninja==1.11.1.1
numpy==1.26.0
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.4.127
opencv-python==4.10.0.84
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.1.2
ordered-set==4.1.0
oss2==2.17.0
packaging==24.2
pandas==2.2.3
pillow==11.0.0
platformdirs==4.3.6
propcache==0.2.0
protobuf==5.28.3
psutil==6.1.0
pyarrow==18.0.0
pycocotools==2.0.8
pycparser==2.22
pycryptodome==3.21.0
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2023.4
PyYAML==6.0.2
regex==2024.11.6
requests==2.28.2
rich==13.4.2
safetensors==0.4.5
scikit-image==0.24.0
scipy==1.13.1
sentry-sdk==2.18.0
setproctitle==1.3.3
shapely==2.0.6
six==1.16.0
smmap==5.0.1
sympy==1.13.1
tabulate==0.9.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
termcolor==2.5.0
terminaltables==3.1.10
tifffile==2024.8.30
tokenizers==0.20.3
tomli==2.1.0
torch==2.1.0+cu121
torchaudio==2.1.0+cu121
torchvision==0.16.0+cu121
tqdm==4.65.2
transformers==4.46.2
triton==2.1.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==1.26.20
wandb==0.18.6
wcwidth==0.2.13
Werkzeug==3.1.3
xxhash==3.5.0
yapf==0.40.2
yarl==1.17.1
zipp==3.21.0
